Embed an absolute path to libdarknet.so in darknet.node
node-gyp (or the build chain it uses) seems to embed only a name of a shared library to a binrary referring to it. This means that the dynamic loader needs to find the shared library in one of the standard locations, which isn't the case here.

I think there should be a way to configure node-gyp to embed an absolute path, but I haven't found it so far. So I made it rewrite it after node-gyp builds darknet.node using install_name_tool.

Note that we can use @rpath as well, but I couldn't find a way to embed @rpath/libdarknet.so into darknet.node neither even though I found a way to set RPATH.

https://blog.s1h.org/shipping-node-addons-with-libraries/